Change pixels of picture to u8 to speed up pipe data to VLC. Fix motion compensation round error

+++ include/picture.h
@@ -15,7 +15,7 @@ class Picture {
         void decode (Stream* vs);
 
         // Essentail info
-        i16* pixel[3];
+        u8*  pixel[3];
         u16  horz_size;
         u16  vert_size;
         u8*  intra_q;
